Refactor style inclusion, move components into dirs
