--- src/App.css
@@ -1,69 +1,0 @@
-/* basic colors, margins, etc. */
-body {
-  background-color: #ffffff;
-  margin: 0;
-}
-
-h2 {
-  color: #000000;
-  text-align: center;
-}
-
-header, section {
-  padding-bottom: 2em;
-}
-
-img {
-  width: 100%;
-}
-
-/* grid styling */
-.basic-grid {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-}
-
-/* form styling */
-form {
-  /* vertically centered */
-  display: flex;
-  flex-direction: column;
-  margin: auto;
-}
-
-/* header styling */
-header {
-  height: 100vh;
-}
-header img {
-  width: 100%;
-  height: 70vh;
-  display: flex;
-}
-
-/* pattern styling
-* sourced from: https://css-pattern.com/ 
-*/
-#background_pattern {
-  height: 30em;
-  --s: 140px; /* control the size */
-  --_g: #0000 52%,#170409 /* first color */ 54% 57%,#0000 59%;
-  background: radial-gradient(farthest-side at -33.33% 50%, var(--_g)) 0 calc(var(--s) / 2), radial-gradient(farthest-side at 50% 133.33%, var(--_g)) calc(var(--s) / 2) 0, radial-gradient(farthest-side at 133.33% 50%, var(--_g)), radial-gradient(farthest-side at 50% -33.33%, var(--_g)), #f1cd01; /* second color */
-  background-size: calc(var(--s) / 4.667) var(--s), var(--s) calc(var(--s) / 4.667);
-}
-
-/* planner styling */
-#planner .basic-grid .map {
-  height: 30em;
-}
-
-/* footer styling */
-/* media queries */
-@media (max-width: 768px) {
-  /* display each element in its own row */
-  .basic-grid {
-    grid-template-columns: 1fr;
-  }
-}
-
-/*# sourceMappingURL=App.css.map */
