+++ src/components/App/App.tsx
@@ -1,10 +1,10 @@
-import './App.css'
-import Card from './components/Card'
-import DestinationContainer from './components/DestinationContainer'
-import InputForm from './components/InputForm'
-import { DestinationInfo } from './components/DestinationContainer'
-import att from './assets/att_disc2.jpg'
-import skyline_art from './assets/skyline_art.png'
+import './App.scss';
+import Card from '../Card/Card.tsx';
+import DestinationContainer from '../DestinationContainer/DestinationContainer.tsx';
+import InputForm from '../InputForm/InputForm.tsx';
+import { DestinationInfo } from '../DestinationContainer/DestinationContainer.tsx';
+import att from '../../assets/att_disc2.jpg';
+import skyline_art from '../../assets/skyline_art.png';
 
 function App() {
   const destinations:DestinationInfo[] = new Array(
