--- src/components/css/Card.css
@@ -1,18 +1,0 @@
-/* card styling */
-.card {
-  display: flex;
-  margin: auto;
-  flex-direction: column;
-}
-.card p {
-  text-align: center;
-  max-width: 30em;
-  margin: auto;
-}
-.card button {
-  width: 10em;
-  margin: auto;
-  margin-bottom: 2em;
-}
-
-/*# sourceMappingURL=Card.css.map */
