--- src/components/css/Destination.css
@@ -1,38 +1,0 @@
-.destination-item {
-  width: 80%;
-  height: 20em;
-  border: #f1cd01 solid 0.5em;
-  display: flex;
-  flex-direction: column;
-  margin: auto;
-  transition: scale 1s;
-}
-.destination-item #logo, .destination-item svg, .destination-item p, .destination-item img {
-  display: flex;
-  margin: auto;
-}
-.destination-item p {
-  text-align: center;
-}
-.destination-item img {
-  height: 90%;
-  width: 80%;
-}
-
-.destination-item:hover {
-  scale: 1.1;
-}
-
-.purple {
-  border: #604675 solid 0.5em;
-}
-
-/* media queries */
-@media (max-width: 768px) {
-  /* display each element in its own row */
-  .destination-item {
-    width: 100%;
-  }
-}
-
-/*# sourceMappingURL=Destination.css.map */
